+++ src/Pim/Bundle/CatalogBundle/AttributeType/NumberType.php
@@ -50,10 +50,6 @@ class NumberType extends AbstractAttributeType
                 'options'   => [
                     'attr' => $attribute->getId() ? [] : ['checked' => 'checked']
                 ]
-            ],
-            'searchable' => [
-                'name'      => 'searchable',
-                'fieldType' => 'switch'
             ]
         ];