+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/ProductQueryBuilder.php
@@ -182,9 +182,11 @@ class ProductQueryBuilder implements ProductQueryBuilderInterface
      */
     public function addFieldSorter($field, $direction)
     {
+        $field = (strpos($field, 'in_group_') !== false) ? 'in_group' : $field;
         $customSorters = [
             'family'       => 'Pim\Bundle\CatalogBundle\Doctrine\ORM\Sorter\FamilySorter',
-            'completeness' => 'Pim\Bundle\CatalogBundle\Doctrine\ORM\Sorter\CompletenessSorter'
+            'completeness' => 'Pim\Bundle\CatalogBundle\Doctrine\ORM\Sorter\CompletenessSorter',
+            'in_group'     => 'Pim\Bundle\CatalogBundle\Doctrine\ORM\Sorter\InGroupSorter'
         ];
 
         if (isset($customSorters[$field])) {