+++ src/Pim/Bundle/InstallerBundle/Command/DatabaseCommand.php
@@ -16,7 +16,7 @@ use Pim\Bundle\CatalogBundle\DependencyInjection\PimCatalogExtension;
  * - creates database
  * - updates schema
  * - loads fixtures
- * - launches other command for database calculations (search index, completeness calculation)
+ * - launches other command for database calculations (completeness calculation)
  *
  * @author    Romain Monceau <romain@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
@@ -124,7 +124,9 @@ class DatabaseCommand extends ContainerAwareCommand
             $input->setOption('fixtures', self::LOAD_ORO);
         }
 
-        $output->writeln('<info>Load fixtures.</info>');
+        $output->writeln(
+            sprintf('<info>Load fixtures. (data set: %s)</info>', $this->getContainer()->getParameter('installer_data'))
+        );
 
         $params = array(
                 '--no-interaction' => true,
@@ -192,11 +194,6 @@ class DatabaseCommand extends ContainerAwareCommand
     protected function launchCommands(InputInterface $input, OutputInterface $output)
     {
         $this->commandExecutor
-            ->runCommand('oro:search:create-index')
-            ->runCommand(
-                'pim:search:reindex',
-                array('locale' => $this->getContainer()->getParameter('locale'))
-            )
             ->runCommand('pim:versioning:refresh')
             ->runCommand('pim:completeness:calculate');